Update: Ship loading in explore has started

+++ Assets/Scripts/Explore/LocationProfiles.cs
@@ -10,26 +10,26 @@ public class LocationProfile
     public string mode;
     public float piracyRisk;
     public float enemyRisk;
-    public float largeCapitalShips;
-    public float mediumCapitalShips;
-    public float smallCapitalShips;
-    public float freighters;
-    public float lightfreighters;
-    public float shuttles;
-    public float fighters;
-    public float stations;
-    public float cargofields;
-    public float navbuoys;
-    public float enemyLargeCapitalShips;
-    public float enemyMediumCapitalShips;
-    public float enemySmallCapitalShips;
-    public float enemyFreighters;
-    public float enemyLightfreighters;
-    public float enemyShuttles;
-    public float enemyFighters;
-    public float pirateLightfreighters;
-    public float pirateShuttles;
-    public float pirateFighters;
+    public int largeCapitalShips;
+    public int mediumCapitalShips;
+    public int smallCapitalShips;
+    public int freighters;
+    public int lightfreighters;
+    public int shuttles;
+    public int fighters;
+    public int stations;
+    public int cargofields;
+    public int navbuoys;
+    public int enemyLargeCapitalShips;
+    public int enemyMediumCapitalShips;
+    public int enemySmallCapitalShips;
+    public int enemyFreighters;
+    public int enemyLightfreighters;
+    public int enemyShuttles;
+    public int enemyFighters;
+    public int pirateLightfreighters;
+    public int pirateShuttles;
+    public int pirateFighters;
 }
 
 [System.Serializable]
